Update P2FunctionsAreUs.py
I created function 2 and 4.  Function 2 allows the end user to choose if they would like to start a new season (choose a home and away teams, randomly and randomly generate scores), view the final record of a previous season, or quit the program.

Function 4 takes the teams selected in function 4, randomly generates game scores and returns a win or a loss.